pyproject.toml
[project]
name = "study-serverless-webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "google-genai>=1.40.0",
    "uvicorn>=0.37.0",
]
